feat: negotiate content type
